+++ app/src/main/java/de/mkalb/etpetssim/engine/model/SimulationStepRunner.java
@@ -17,6 +17,6 @@ public interface SimulationStepRunner<C> {
      * @param stepIndex the index of the current simulation step
      * @param context   the context object used to share or accumulate state during the simulation
      */
-    void performStep(long stepIndex, C context);
+    void performStep(int stepIndex, C context);
 
 }
